% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_example_data_path.R
\name{gm_example_data_path}
\alias{gm_example_data_path}
\title{Obter caminho de arquivos de exemplo do pacote}
\usage{
gm_example_data_path(file = NULL)
}
\arguments{
\item{file}{Nome do arquivo dentro de \code{extdata}. Se \code{NULL}, retorna
o caminho da pasta \code{extdata}.}
}
\value{
Um vetor de caracteres com o caminho completo para o arquivo ou
diretório solicitado.
}
\description{
Retorna o caminho para os dados de exemplo incluídos no pacote, localizados
na pasta \code{inst/extdata}. Esses dados são utilizados nas aulas e exemplos
do pacote.
}
\examples{
\dontrun{
# Listar arquivos disponíveis
list.files(gm_example_data_path())

# Obter caminho de um arquivo específico
setores_path <- gm_example_data_path("setores.gpkg")
}

}
