% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_make_grid.R
\name{gm_make_grid}
\alias{gm_make_grid}
\title{Criar grade regular}
\usage{
gm_make_grid(x, cellsize = 500)
}
\arguments{
\item{x}{objeto sf}

\item{cellsize}{tamanho da célula em metros}
}
\value{
objeto sf com grid
}
\description{
Criar grade regular
}
