% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mediana_idade.R
\name{gm_mediana_idade}
\alias{gm_mediana_idade}
\title{Calcula a idade mediana aproximada por linha a partir de faixas etárias}
\usage{
gm_mediana_idade(
  df,
  colunas = c("pop_abs_00_04", "pop_abs_05_09", "pop_abs_10_14", "pop_abs_15_19",
    "pop_abs_20_24", "pop_abs_25_29", "pop_abs_30_39", "pop_abs_40_49", "pop_abs_50_59",
    "pop_abs_60_69", "pop_abs_70_mais"),
  breaks_inf = c(0, 5, 10, 15, 20, 25, 30, 40, 50, 60, 70),
  breaks_sup = c(5, 10, 15, 20, 25, 30, 40, 50, 60, 70, 100),
  na_strategy = c("zero", "ignore")
)
}
\arguments{
\item{df}{data frame com as colunas das faixas etárias.}

\item{colunas}{vetor com os nomes das colunas de faixas etárias.}

\item{breaks_inf}{vetor com os limites inferiores das faixas.}

\item{breaks_sup}{vetor com os limites superiores das faixas.}

\item{na_strategy}{estratégia para tratamento de NA:
"zero" substitui NA por 0;
"ignore" retorna NA para a linha se houver qualquer NA.}
}
\value{
vetor numérico com a idade mediana aproximada por linha.
}
\description{
Calcula a idade mediana aproximada por linha a partir de faixas etárias
}
