% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_plot_basic.R
\name{gm_plot_basic}
\alias{gm_plot_basic}
\title{Mapa básico de setores}
\usage{
gm_plot_basic(x, var, titulo = NULL, legenda = NULL)
}
\arguments{
\item{x}{objeto sf.}

\item{var}{nome da variável a ser mapeada.}

\item{titulo}{título do mapa. Se NULL, usa o nome da variável.}

\item{legenda}{título da legenda. Se NULL, usa o nome da variável.}
}
\value{
objeto tmap.
}
\description{
Cria um mapa temático simples a partir de um objeto sf.
}
