% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_example_cycleways.R
\name{gm_read_example_cycleways}
\alias{gm_read_example_cycleways}
\title{Ler ciclovias de exemplo do OpenStreetMap}
\usage{
gm_read_example_cycleways()
}
\value{
Objeto \code{sf} contendo linhas representando ciclovias.
}
\description{
Carrega dados de infraestrutura cicloviária previamente preparados e
incluídos no pacote para uso em aula. Os dados foram obtidos do
OpenStreetMap em Campos dos Gotytacazes
}
\examples{
\dontrun{
ciclovias <- gm_read_example_cycleways()

plot(sf::st_geometry(ciclovias), col = "blue")

}
}
