% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_example_roads.R
\name{gm_read_example_roads}
\alias{gm_read_example_roads}
\title{Ler vias de exemplo}
\usage{
gm_read_example_roads()
}
\value{
objeto sf com vias principais.
}
\description{
Ler vias de exemplo
}
