% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_example_sectors.R
\name{gm_read_example_sectors}
\alias{gm_read_example_sectors}
\title{Ler setores de exemplo de Campos}
\usage{
gm_read_example_sectors()
}
\value{
Objeto \code{sf} contendo linhas representando os setores censitários
}
\description{
Carrega a shape dos setores censitários incluídos no pacote para uso em aula.
Os dados são do Censo 2022 para Campos dos Gotytacazes
}
\examples{
\dontrun{
setores <- gm_read_example_sectors()

plot(sf::st_geometry(setores), col = "blue")

}
}
