% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_example_data.R
\name{gm_example_data}
\alias{gm_example_data}
\title{Retorna o caminho de arquivos de exemplo do pacote}
\usage{
gm_example_data(file = NULL)
}
\arguments{
\item{file}{nome do arquivo em extdata}
}
\description{
Retorna o caminho de arquivos de exemplo do pacote
}
