% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_join_data.R
\name{gm_join_data}
\alias{gm_join_data}
\title{Juntar setores e atributos}
\usage{
gm_join_data(
  setores,
  dados,
  id_setor_geom = "CD_SETOR",
  id_setor_tab = "CD_SETOR"
)
}
\arguments{
\item{setores}{objeto sf com geometria}

\item{dados}{tabela de atributos}

\item{id_setor_geom}{nome da coluna ID no objeto espacial}

\item{id_setor_tab}{nome da coluna ID na tabela}
}
\value{
objeto sf
}
\description{
Juntar setores e atributos
}
