% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_make_grid.R
\name{gm_make_grid}
\alias{gm_make_grid}
\title{Criar grade regular}
\usage{
gm_make_grid(
  shape,
  type = c("metric", "latlon"),
  cellsize = 100,
  crs_out = NULL,
  clip = TRUE
)
}
\arguments{
\item{shape}{Objeto \code{sf}, como a malha de setores censitários.}

\item{type}{Tipo de grade. Use \code{"metric"} para grade em metros ou \code{"latlon"}
para grade em graus.}

\item{cellsize}{Tamanho da célula. Quando \code{type = "metric"}, deve ser um
valor numérico em metros. Quando \code{type = "latlon"}, pode ser um valor
numérico em graus ou um dos valores: \code{"1d"}, \code{"30m"}, \code{"15m"}, \code{"1m"},
\code{"30s"} ou \code{"15s"}.}

\item{crs_out}{Sistema de coordenadas usado para criar a grade. Se \code{NULL},
usa \code{31984} para \code{type = "metric"} e \code{4674} para \code{type = "latlon"}.}

\item{clip}{Se \code{TRUE}, mantém apenas as células que intersectam a geometria
de \code{shape}.}
}
\value{
Objeto \code{sf} com a grade regular.
}
\description{
Cria uma grade regular a partir da extensão espacial de um objeto \code{sf}.
}
\details{
A grade pode ser criada em coordenadas métricas, como UTM, ou em coordenadas
geográficas, alinhada à latitude e longitude.
}
\examples{
\dontrun{
base <- gm_join_data(
  gm_read_sectors(gm_example_data("setores.gpkg")),
  gm_read_attributes(gm_example_data("variaveis.rds"))
)

grid_utm <- gm_make_grid(base, type = "metric", cellsize = 1000)
grid_latlon <- gm_make_grid(base, type = "latlon", cellsize = "15s")
}

}
