% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_osm_roads.R
\name{gm_osm_roads}
\alias{gm_osm_roads}
\title{Baixar vias do OpenStreetMap}
\usage{
gm_osm_roads(
  shape,
  highway = c("motorway", "trunk", "primary", "secondary", "tertiary"),
  timeout = 180
)
}
\arguments{
\item{shape}{objeto sf usado para definir a área de busca.}

\item{highway}{tipos de vias do OSM.}

\item{timeout}{tempo máximo da consulta, em segundos.}
}
\value{
objeto sf com vias.
}
\description{
Baixar vias do OpenStreetMap
}
