% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_plot_basic.R
\name{gm_plot_basic}
\alias{gm_plot_basic}
\title{Mapa básico de setores}
\usage{
gm_plot_basic(x, var, titulo = NULL)
}
\arguments{
\item{x}{objeto sf}

\item{var}{nome da variável}

\item{titulo}{título do mapa}
}
\value{
ggplot
}
\description{
Mapa básico de setores
}
