% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_plot_overlay.R
\name{gm_plot_overlay}
\alias{gm_plot_overlay}
\title{Mapa com duas camadas espaciais}
\usage{
gm_plot_overlay(
  base,
  overlay,
  titulo = NULL,
  base_fill = "grey90",
  base_col = "white",
  base_alpha = 0.7,
  overlay_fill = NA,
  overlay_col = "red",
  overlay_alpha = 0.7,
  overlay_lwd = 2
)
}
\arguments{
\item{base}{objeto sf da camada de base, por exemplo setores censitários.}

\item{overlay}{objeto sf da camada sobreposta, por exemplo uma grade regular.}

\item{titulo}{título do mapa.}

\item{base_fill}{cor de preenchimento da camada de base.}

\item{base_col}{cor da borda da camada de base.}

\item{overlay_fill}{cor de preenchimento da camada sobreposta.}

\item{overlay_col}{cor da borda da camada sobreposta.}

\item{overlay_alpha}{transparência da camada sobreposta.}

\item{overlay_lwd}{espessura da camada sobreposta}
}
\value{
objeto tmap.
}
\description{
Cria um mapa simples com uma camada de base e uma camada sobreposta.
}
