% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_attributes.R
\name{gm_read_attributes}
\alias{gm_read_attributes}
\title{Ler atributos de setores censitários}
\usage{
gm_read_attributes(path)
}
\arguments{
\item{path}{caminho para csv local.}
}
\value{
data.frame
}
\description{
Ler atributos de setores censitários
}
