% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_example_cycleways.R
\name{gm_read_example_cycleways}
\alias{gm_read_example_cycleways}
\title{Ler ciclovias de exemplo do OpenStreetMap}
\usage{
gm_read_example_cycleways(type = c("cycleway", "cycleway_tag"))
}
\arguments{
\item{type}{Tipo de dado a ser carregado. Deve ser um dos seguintes:
\code{"cycleway"} ou \code{"cycleway_tag"}.}
}
\value{
Objeto \code{sf} contendo linhas representando ciclovias.
}
\description{
Carrega dados de infraestrutura cicloviária previamente preparados e
incluídos no pacote para uso em aula. Os dados foram obtidos do
OpenStreetMap e representam duas formas distintas de modelagem de
ciclovias:
}
\details{
\itemize{
\item \code{"cycleway"}: vias exclusivas para bicicletas
(tag \code{highway = cycleway})
\item \code{"cycleway_tag"}: ciclovias associadas a vias existentes
(tag \code{cycleway = *})
}

Essa distinção é importante, pois a infraestrutura cicloviária pode
estar representada tanto como uma geometria própria quanto como um
atributo de uma via.
}
\examples{
\dontrun{
ciclovias <- gm_read_example_cycleways("cycleway")
ciclofaixas <- gm_read_example_cycleways("cycleway_tag")

plot(sf::st_geometry(ciclovias), col = "blue")
plot(sf::st_geometry(ciclofaixas), add = TRUE, col = "green")
}

}
