% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_neighborhoods.R
\name{gm_read_neighborhoods}
\alias{gm_read_neighborhoods}
\title{Ler bairros de um município}
\usage{
gm_read_neighborhoods(code_muni)
}
\arguments{
\item{code_muni}{código IBGE do município}
}
\value{
objeto sf
}
\description{
Ler bairros de um município
}
