% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_sectors.R
\name{gm_read_sectors}
\alias{gm_read_sectors}
\title{Ler setores censitários}
\usage{
gm_read_sectors(path = NULL, code_muni = NULL, year = 2022)
}
\arguments{
\item{path}{caminho para geopackage/shapefile dos setores.}

\item{code_muni}{código IBGE do município, opcional.}

\item{year}{ano da malha.}
}
\value{
objeto sf
}
\description{
Ler setores censitários
}
