% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_read_attributes.R
\name{gm_read_tracts_censo}
\alias{gm_read_tracts_censo}
\title{Ler dados de setores do Censo via censobr}
\usage{
gm_read_tracts_censo(dataset = "Basico", year = 2022, as_data_frame = TRUE)
}
\arguments{
\item{dataset}{base do censo}

\item{year}{ano do censo}

\item{as_data_frame}{retornar data.frame}
}
\description{
Ler dados de setores do Censo via censobr
}
